requirements.txt
python-telegram-bot[webhooks]==21.10
httpx==0.27.2
python-dotenv==1.0.1
uvicorn==0.34.0
fastapi==0.115.8
gunicorn
supabase
openai
anthropic
google-genai>=1.0;
requests
email-validator>=2.2;
pydantic[email]>=2.12;
sqlalchemy[asyncio]
google-generativeai
asyncpg
slowapi
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
stripe>=10.0.0;
